Now passing refactored test cases

+++ src/main/java/au/edu/wehi/idsv/UnmappedMateReadPair.java
@@ -12,11 +12,11 @@ public class UnmappedMateReadPair extends NonReferenceReadPair {
 		return String.format("UM %s MQ=%d RN=%s", getBreakendSummary(), getLocalMapq(), getLocalledMappedRead().getReadName());
 	}
 	@Override
-	public double getBreakendQual() {
+	public float getBreakendQual() {
 		IdsvMetrics metrics = getEvidenceSource().getMetrics().getIdsvMetrics();
 		// completely unmapped read pairs are excluded for consistency with sc and dp calculation
 		double score = -10 * Math.log10((double)metrics.READ_PAIRS_ONE_MAPPED / (double)(metrics.READ_PAIRS - metrics.READ_PAIRS_ZERO_MAPPED));
 		score = Math.min(score, getLocalMapq());
-		return score;
+		return (float)score;
 	}
 }
